#1088 - Added AreaType to Location, renamed receptionDate to districtLevelDate, added regionLevelDate and nationalLevelDate, added country column to detailed case export, enabled vaccination doses for yellow fever, added "associated with outbreak" to detailed case export

+++ sormas-backend/src/main/java/de/symeda/sormas/backend/region/DistrictFacadeEjb.java
@@ -192,7 +192,7 @@ public class DistrictFacadeEjb implements DistrictFacade {
 	}
 
 	@Override
-	public DistrictReferenceDto getDistrictReferenceById(int id) {
+	public DistrictReferenceDto getDistrictReferenceById(long id) {
 		return toReferenceDto(districtService.getById(id));
 	}
